Create a Path from Components

path_create() creates a cleaned path from the components specified in ... and an optional file extension (ext). The path is automatically expanded using either the R or system definition of the home directory.

path_create(..., ext = NULL, home = c("r", "user"))

Arguments

...
character vectors, if any values are NA, the result will also be NA. The paths follow the recycling rules used in the tibble package, namely that only length 1 arguments are recycled.

ext
An optional extension to append to the generated path.

home
The home directory definition to use when expanding a path. The default is to use R's definition, but the system definition of the user's home directory can also be used. These are equivalent to calling path_expand_r() or path_expand(), respectively.

Value

The combined and cleaned path in an fs_path object, which is a character vector that also has class fs_path